Refactor Player into its own module
Seeing the outline of some traits here: "renderable" and "udpateable".
This tidies up main.rs quite a bit for now, and ensures the Player object
owns its own texture. Maybe putting the render logic in the Player isn't
the best idea (what if we want to change renderer to e.g. winit and OpenGL
or some kind of tty renderer?). Might try to fix that next.