Try to decouple Player from the rendering details
Add a Renderable trait. This trait is generic over the rendering target.
This means we can reimplement it for a different target if we want to
render to something other than an SDL2 canvas. However, this is still
hard-coded in the Player module. We still have to have all of these
Renderable implementations on the Player type. This includes having to
import all of the SDL types into the Player module. I'm still trying to
figure out the best way to tidy this up. I think it's basically a case
of moving the implementation of the Renderable trait to a different
module.