Refactor to separate rendering from logic
Not 100% perfect. I think it won't be until the Player type doesn't know
about specific texture types. However, this was an interesting exercise
in trying to move the knowledge of rendering the game away from the
types responsible for the internal logic. In theory we could substitute
out a different renderer with a different target (e.g. a tty) and not
have to change the Player type at all. Obviously, at the moment the
Player still knows about sprites, textures and the like, so there is
still work to be done here.

I had to make use of the unsafe_textures feature of the sdl2 crate to
get this working. What this means is that there is some
chance of the Texture outliving the TextureCreator that made it. I am
not sure what the implications of this are in practice, but it will do
for this little playground for now.