Separate any knowledge of SDL from logical model
Move all of the SDL rendering into a render_sdl package. Only this
package and main.rs now contain any imports for SDL2 specific packages.
This theoretically makes it possible to swap out other renderers as
needed. I am making use of the Deref and DerefMut traits to pass
requests through the sdl_player struct to the underlying player.
Unfortunately I got type woes when trying to use this with the update
function, since it returns a Player, and I need an SdlPlayer type.
Overriding the update function to return an SdlPlayer solves this, but I
can't help but feel like there's a nicer way of dealing with all of
this.